add format data file, add temp buttons and handlers for modals
